refactor: rename for consistency
